#104: fix test failures (jdbc Short→Integer, JWT sub UUID, FK cascade) + regen openapi snapshot
Three fixes to land green:
- V8 migration: drop the conditional GRANT/REVOKE on alpenflight_app —
  forbidden-migration-patterns bans bare GRANT/REVOKE tokens; the
  append-only-via-role defense becomes a follow-up ops story alongside
  the migrator/app role split.
- audit ITs: pg JDBC maps SMALLINT to Integer in queryForList; the
  http_status assertions now compare against int, not short.
- ClubsAuditEventIT: mint the JWT with an explicit UUID sub so
  ActorResolver can resolve it to a non-null actor_keycloak_sub.
- MemberStateTenantIsolationIT (boyscout): also DELETE location +
  inoutbound_point rows in cleanupPreviousRun — LocationsAuthorizationIT
  shares the same CLUB_A/CLUB_B UUIDs and may leave location rows that
  trip fk_location_club_id on the trailing club delete.
- Regenerate web/openapi/openapi.json to include the new
  GET /api/v1/admin/audit-events surface.

+++ alpenflight/server/src/test/java/ch/alpenflight/audit/web/AuditFailedRequestIT.java
@@ -75,7 +75,8 @@ class AuditFailedRequestIT extends PostgresIntegrationTest {
         Map<String, Object> row = failedRows.get(0);
         assertThat(row.get("action")).isEqualTo("CREATE");
         assertThat(row.get("target_entity_type")).isEqualTo("Club");
-        assertThat(row.get("http_status")).isEqualTo((short) 400);
+        // pg JDBC maps SMALLINT to Integer in queryForList default mapping
+        assertThat(row.get("http_status")).isEqualTo(400);
         assertThat(row.get("after_state")).isNull();
         assertThat(row.get("failure_reason")).isNotNull();
     }
